Make canonicalURL memo test a server-only realm-server test
runSharedTest is for isomorphic tests that run in both the server and the
browser. This test only exercises the index-runner canonicalURL memo, which
is server-side, so it does not need the shared-test indirection. Inline the
stub network and assertions directly in the realm-server test file and drop
the runtime-common shared module.
